modulization: separate DUAnalyzer to another file
